Move ServerError to shared and add signup as backend feature
