+++ client/src/services/assignmentServices.ts
@@ -1,7 +1,7 @@
 import { AxiosResponse, AxiosError } from 'axios';
 import { api } from 'client/utils/api';
 import { PostPlayerAssignmentResponse } from 'client/typings/result.typings';
-import { ServerError } from 'client/typings/utils.typings';
+import { ServerError } from 'shared/typings/api/errors';
 import { ASSIGNMENT_ENDPOINT } from 'shared/constants/apiEndpoints';
 
 export const postPlayerAssignment = async (
