+++ client/src/services/groupServices.ts
@@ -5,7 +5,7 @@ import {
   PostGroupResponse,
   GetGroupResponse,
 } from 'client/typings/group.typings';
-import { ServerError } from 'client/typings/utils.typings';
+import { ServerError } from 'shared/typings/api/errors';
 import { GROUP_ENDPOINT } from 'shared/constants/apiEndpoints';
 
 export const postGroup = async (
