+++ client/src/services/resultsServices.ts
@@ -1,6 +1,6 @@
 import { AxiosResponse, AxiosError } from 'axios';
 import { api } from 'client/utils/api';
-import { ServerError } from 'client/typings/utils.typings';
+import { ServerError } from 'shared/typings/api/errors';
 import { GetResultsResponse } from 'client/typings/result.typings';
 import { RESULTS_ENDPOINT } from 'shared/constants/apiEndpoints';
