+++ shared/typings/api/errors.ts
@@ -1,0 +1,5 @@
+export interface ServerError {
+  code: number;
+  message: string;
+  status: 'error';
+}
